refactor _common_kwargs to OmegaRuntime
