+++ omegaml/defaults.py
@@ -28,7 +28,9 @@ OMEGA_NOTEBOOK_COLLECTION = 'ipynb'
 OMEGA_RESULT_BACKEND = 'amqp'
 #: the celery configurations
 OMEGA_CELERY_CONFIG = {
-    'CELERY_ACCEPT_CONTENT': ['pickle', 'json', 'msgpack', 'yaml'],
+    'CELERY_ACCEPT_CONTENT': ['pickle', 'json'],
+    'CELERY_TASK_SERIALIZER': 'pickle',
+    'CELERY_RESULT_SERIALIZER': 'pickle',
     'BROKER_URL': OMEGA_BROKER,
     'CELERY_RESULT_BACKEND': OMEGA_RESULT_BACKEND,
     'CELERY_ALWAYS_EAGER': False if 'OMEGA_BROKER' in os.environ else True,
