+++ omegaml/runtimes/mixins/modelmixin.py
@@ -11,10 +11,6 @@ logger = logging.getLogger(__file__)
 
 
 class ModelMixin(object):
-    @property
-    def _common_kwargs(self):
-        return dict(pure_python=self.pure_python)
-
     def fit(self, Xname, Yname=None, **kwargs):
         """
         fit the model
@@ -36,7 +32,7 @@ class ModelMixin(object):
         if Yname is not None:
             Yname = self._ensure_data_is_stored(Yname, prefix='_fitY')
         return omega_fit.delay(self.modelname, Xname, Yname,
-                               **self._common_kwargs, **kwargs)
+                               **self.runtime._common_kwargs, **kwargs)
 
     def partial_fit(self, Xname, Yname=None, **kwargs):
         """
@@ -59,7 +55,7 @@ class ModelMixin(object):
         if Yname is not None:
             Yname = self._ensure_data_is_stored(Yname, prefix='_fitY')
         return omega_fit.delay(self.modelname, Xname, Yname,
-                               **self._common_kwargs, **kwargs)
+                               **self.runtime._common_kwargs, **kwargs)
 
     def transform(self, Xname, rName=None, **kwargs):
         """
@@ -77,7 +73,7 @@ class ModelMixin(object):
         Xname = self._ensure_data_is_stored(Xname)
         return omega_transform.delay(self.modelname, Xname,
                                      rName=rName,
-                                     **self._common_kwargs, **kwargs)
+                                     **self.runtime._common_kwargs, **kwargs)
 
     def fit_transform(self, Xname, Yname=None, rName=None, **kwargs):
         """
@@ -100,7 +96,7 @@ class ModelMixin(object):
             Yname = self._ensure_data_is_stored(Yname)
         return omega_fit_transform.delay(self.modelname, Xname, Yname,
                                          rName=rName, transform=True,
-                                         **self._common_kwargs, **kwargs)
+                                         **self.runtime._common_kwargs, **kwargs)
 
     def predict(self, Xpath_or_data, rName=None, **kwargs):
         """
@@ -117,7 +113,7 @@ class ModelMixin(object):
         omega_predict = self.runtime.task('omegaml.tasks.omega_predict')
         Xname = self._ensure_data_is_stored(Xpath_or_data)
         return omega_predict.delay(self.modelname, Xname, rName=rName,
-                                   **self._common_kwargs, **kwargs)
+                                   **self.runtime._common_kwargs, **kwargs)
 
     def predict_proba(self, Xpath_or_data, rName=None, **kwargs):
         """
@@ -135,7 +131,7 @@ class ModelMixin(object):
             'omegaml.tasks.omega_predict_proba')
         Xname = self._ensure_data_is_stored(Xpath_or_data)
         return omega_predict_proba.delay(self.modelname, Xname, rName=rName,
-                                         **self._common_kwargs, **kwargs)
+                                         **self.runtime._common_kwargs, **kwargs)
 
     def score(self, Xname, yName, rName=None, **kwargs):
         """
@@ -154,7 +150,7 @@ class ModelMixin(object):
         Xname = self._ensure_data_is_stored(Xname)
         yName = self._ensure_data_is_stored(yName)
         return omega_score.delay(self.modelname, Xname, yName, rName=rName,
-                                 **self._common_kwargs, **kwargs)
+                                 **self.runtime._common_kwargs, **kwargs)
 
     def decision_function(self, Xname, rName=None, **kwargs):
         """
@@ -171,7 +167,7 @@ class ModelMixin(object):
         omega_decision_function = self.runtime.task('omegaml.tasks.omega_decision_function')
         Xname = self._ensure_data_is_stored(Xname)
         return omega_decision_function.delay(self.modelname, Xname, rName=rName,
-                                             **self._common_kwargs, **kwargs)
+                                             **self.runtime._common_kwargs, **kwargs)
 
     def _ensure_data_is_stored(self, name_or_data, prefix='_temp'):
         if is_dataframe(name_or_data):
